[JT-60] refactor: 리팩토링 전 코드 분리
