+++ module-application/src/main/java/com/devtoon/jtoon/security/filter/JwtAuthenticationFilter.java
@@ -2,22 +2,25 @@ package com.devtoon.jtoon.security.filter;
 
 import static com.devtoon.jtoon.security.util.SecurityConstant.*;
 
+import java.io.IOException;
+import java.util.Arrays;
+
 import org.jetbrains.annotations.NotNull;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.filter.OncePerRequestFilter;
 import org.springframework.web.servlet.HandlerExceptionResolver;
 
+import com.devtoon.jtoon.global.common.MemberThreadLocal;
 import com.devtoon.jtoon.security.application.JwtService;
 import com.devtoon.jtoon.security.domain.jwt.CustomUserDetails;
-import com.devtoon.jtoon.security.domain.jwt.MemberThreadLocal;
 import com.devtoon.jtoon.security.util.TokenCookie;
+
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import java.io.IOException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 
@@ -34,7 +37,11 @@ public class JwtAuthenticationFilter extends OncePerRequestFilter {
 		@NotNull HttpServletResponse response,
 		@NotNull FilterChain filterChain
 	) throws ServletException, IOException {
-		String accessToken = request.getHeader(ACCESS_TOKEN_HEADER); //TODO: getCookies로 처리?
+		String accessToken = Arrays.stream(request.getCookies())
+			.filter(coo -> coo.getName().equals(ACCESS_TOKEN_HEADER))
+			.map(Cookie::getValue)
+			.findFirst()
+			.orElse(null);
 
 		if (accessToken != null && accessToken.startsWith(BEARER_VALUE)) {
 			try {
@@ -54,7 +61,12 @@ public class JwtAuthenticationFilter extends OncePerRequestFilter {
 	}
 
 	private String validateAndGetRefreshToken(HttpServletRequest request) {
-		String refreshToken = request.getHeader(REFRESH_TOKEN_HEADER);
+		String refreshToken = Arrays.stream(request.getCookies())
+			.filter(coo -> coo.getName().equals(REFRESH_TOKEN_HEADER))
+			.map(Cookie::getValue)
+			.findFirst()
+			.orElse(null);
+
 		refreshToken = refreshToken.split(SPLIT_DATA)[1];
 		jwtService.isTokenValid(refreshToken);
 		jwtService.verifyRefreshTokenDb(refreshToken);
