+++ module-domain/src/main/java/com/devtoon/jtoon/member/entity/MemberCookie.java
@@ -1,11 +1,10 @@
-package com.devtoon.jtoon.payment.entity;
+package com.devtoon.jtoon.member.entity;
 
 import static com.devtoon.jtoon.error.model.ErrorStatus.*;
 import static java.util.Objects.*;
 
 import com.devtoon.jtoon.error.exception.InvalidRequestException;
 import com.devtoon.jtoon.global.common.BaseTimeEntity;
-import com.devtoon.jtoon.member.entity.Member;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
