+++ module-domain/src/main/java/com/devtoon/jtoon/payment/service/PaymentInfoService.java
@@ -9,8 +9,8 @@ import com.devtoon.jtoon.error.exception.DuplicatedException;
 import com.devtoon.jtoon.error.exception.InvalidRequestException;
 import com.devtoon.jtoon.error.model.ErrorStatus;
 import com.devtoon.jtoon.member.entity.Member;
+import com.devtoon.jtoon.member.entity.MemberCookie;
 import com.devtoon.jtoon.payment.entity.CookieItem;
-import com.devtoon.jtoon.payment.entity.MemberCookie;
 import com.devtoon.jtoon.payment.entity.PaymentInfo;
 import com.devtoon.jtoon.payment.repository.MemberCookieRepository;
 import com.devtoon.jtoon.payment.repository.PaymentInfoRepository;
